Changed Swing Mechanics
I changed the Project Settings by adding a collision channel (with permission) as well as adding an extra socket to the Devvy Skeleton so as to not mess with the original sockets. These changes will not have an affect on other groups' mechanics. Otherwise, new blueprints were added to create the swing mechanics.